完成FTP程序

+++ module3/ftp/core/server.py
@@ -7,7 +7,11 @@ import subprocess
 import os
 import json
 import shutil
-from .ftp_tools import check_auth, hash_password
+import sys
+BASEDIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+sys.path.insert(0, BASEDIR)
+from core.ftp_tools import check_auth, hash_password, record_log, check_root
+from core.read_config import GetConfig
 
 
 class MyTCPHandler(socketserver.BaseRequestHandler):
@@ -15,6 +19,7 @@ class MyTCPHandler(socketserver.BaseRequestHandler):
     BASEDIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
     AUTHDIR = BASEDIR + os.sep + 'conf'
     transfer_count = 4096  # 全局变量，网络传输大小，字节
+    change_dir_error = 'due to can not change directory.'
     comm_dict = {
         'cd': 'Change Dir',
         'dir': 'List Dir',
@@ -24,15 +29,18 @@ class MyTCPHandler(socketserver.BaseRequestHandler):
         'del': 'Delete file or directory',
         'bye': 'Exit.'
     }
-    error_dict = {
-        'comm_error': 'Invalid Commands',
-    }
+    loglevel = GetConfig(os.path.join(AUTHDIR, 'config.ini')).get_config('server')['loglevel']
+    logger = record_log(int(loglevel))
 
     def setup(self):
-        print('%s is connected.' % str(self.client_address))
+        if not self.logger:
+            return False
+        self.logger.info('[{0}] {0} is connected.'.format(str(self.client_address)))
         auth_info = self.request.recv(self.transfer_count).decode('utf-8')  # 接受client端账号密码
         self.account, password = auth_info.split()
         self.password = hash_password(password)     # 加密
+        self.logger.info('[%s] Account: %s' % (str(self.client_address), self.account))
+        # 用户家目录
         auth = check_auth(self.account, self.password, os.path.join(self.AUTHDIR, 'shadow'))    # 鉴权
         if not auth:    # 鉴权失败，发送1
             self.request.sendall(bytes(str(1), encoding='utf-8'))
@@ -43,66 +51,68 @@ class MyTCPHandler(socketserver.BaseRequestHandler):
     def handle(self):
         check = True
         if self.client_sign == 'bye':
-            print('%s auth failed.' % str(self.client_address))
-            self.finish()
+            self.logger.warning('[%s] "%s" Auth failed.' % (str(self.client_address), self.account))
             check = False
         else:
-            print('[%s:%s] is connected.' % self.client_address)
+            self.logger.info('[{0}] "{1}" Auth success.'.format(str(self.client_address), self.account))
         self.request.sendall(bytes('Welcome %s, Please enter "help" to help.' % str(self.client_address), \
                                    encoding='utf-8'))
+        self.basedir = self.BASEDIR + os.sep + 'dirs' + os.sep + self.account
+        os.chdir(self.basedir)  # 切换目录到用户家目录
         while check:
             try:
                 count = 0
                 recv_list = []
                 num = self.request.recv(self.transfer_count)
                 try:
-                    print('Num: ', num)
+                    self.logger.debug('[%s] Recive data length: %s' % (str(self.client_address), num))
                     num = float(num.decode('utf-8'))
                 except ValueError as e:
-                    print(e)
+                    self.logger.error('[%s] %s' % (str(self.client_address), e))
                     break
                 except Exception as e:
-                    print(e)
+                    self.logger.error('[%s] %s' % (str(self.client_address), e))
                     break
                 while count < num:
                     recv_data = self.request.recv(self.transfer_count)
-                    print(recv_data)
                     recv_list.append(recv_data.decode('utf-8'))
                     count += self.transfer_count
                 self.recv = ''.join(recv_list)
-                print('Recv:', self.recv)
+                self.logger.info('[%s] Command: %s' % (str(self.client_address), self.recv))
                 try:
                     action = self.recv.split()[0]
                 except IndexError as e:
-                    print(e)
+                    self.logger.error('[%s] %s' % (str(self.client_address), e))
                     continue
                 if action == 'bye':
                     self.finish()
                 if action not in self.comm_dict:
+                    self.logger.debug('[%s] Invalid command: %s' % (str(self.client_address), self.recv))
                     self.request.sendall(bytes('Invalid Command.', encoding='utf-8'))
                     continue
                 func = getattr(self, '_%s' % action, self._comm)
-                print('_%s' % action)
+                self.logger.debug('[%s] Action: _%s' % (str(self.client_address), action))
                 func()
                 continue
             except (ConnectionResetError, ConnectionAbortedError) as e:
-                print(e)
+                self.logger.error('[%s] %s' % (self.client_address, e))
                 continue
             except FileNotFoundError as e:
-                print(e)
+                self.logger.error('[%s] %s' % (self.client_address, e))
                 self.request.sendall(bytes('Error...[File Not Found]', encoding='utf-8'))
                 continue
 
     def finish(self):
-        print('[%s:%s] is disconnected.' % self.client_address)
+        self.logger.warning('[{0}] {0} is disconnected.'.format(str(self.client_address)))
 
     def _put(self):
-        action, filename, filesize = self.recv.split()[0], self.recv.split()[1], int(self.recv.split()[2])
+        recv_data = self.recv.split()
+        action, filename, filesize = recv_data[0], recv_data[1], recv_data[2]
         self.request.sendall(bytes('Ready...', encoding='utf-8'))
         filecount = 0
         w_file = open(os.path.basename(filename), 'wb')
         while 1:
-            if filecount < filesize:
+            if filecount < int(filesize):
                 recv_data = self.request.recv(self.transfer_count)
                 w_file.write(recv_data)
                 filecount += self.transfer_count
@@ -113,23 +123,39 @@ class MyTCPHandler(socketserver.BaseRequestHandler):
                 break
 
     def _get(self):
-        action, filename = self.recv.split()[0], self.recv.split()[1]
+        recv_data = self.recv.split()
+        action, filename = recv_data[0], recv_data[1]
+        if not check_root(self.basedir, os.path.abspath(filename)):
+            self.request.sendall(bytes('Can not change directory.', encoding='utf-8'))
+            self.logger.warning('[%s] List dir "%s" failed, %s' % \
+                                (str(self.client_address), os.path.abspath(filename), self.change_dir_error))
+            return False
         filesize = os.path.getsize(filename)
         self.request.send(bytes(str(filesize), encoding='utf-8'))
         try:
             with open(filename, 'rb') as f:
                 self.request.sendall(f.read())
-                print(os.path.getsize(filename))
+                self.logger.debug('[%s] File size: %s' % \
+                                  (str(self.client_address), os.path.getsize(filename)))
         except FileNotFoundError as e:
-            print(e)
+            self.logger.error('[%s] %s' % (str(self.client_address), e))
 
     def _help(self):
         send_data = json.dumps(self.comm_dict)
         self.request.sendall(bytes(send_data, encoding='utf-8'))
 
-    def _cd(self, recv_data):
+    def _cd(self):
         e = ''
-        action, dirname = recv_data
+        recv_data = self.recv.split()
+        if len(recv_data) > 1:
+            action, dirname = recv_data[0], recv_data[1]
+        else:
+            action, dirname = recv_data[0], '.'
+        if not check_root(self.basedir, os.path.abspath(dirname)):
+            self.request.sendall(bytes('Can not change directory.', encoding='utf-8'))
+            self.logger.warning('[%s] Change dir "%s" failed, %s' % \
+                                (str(self.client_address), os.path.abspath(dirname), self.change_dir_error))
+            return False
         try:
             subprocess.os.chdir(dirname)
             self.request.sendall(bytes('Operate Success.', encoding='utf-8'))
@@ -138,22 +164,41 @@ class MyTCPHandler(socketserver.BaseRequestHandler):
 
     def _dir(self):
         dirname = '.'
-        if len(self.recv.split()) > 1:
-            dirname = self.recv.split()[1]
+        recv_data = self.recv.split()
+        if len(recv_data) > 1:
+            dirname = recv_data[1]
+        if not check_root(self.basedir, os.path.abspath(dirname)):
+            self.request.sendall(bytes('Can not change directory.', encoding='utf-8'))
+            self.logger.warning('[%s] List dir "%s" failed, %s' % \
+                                (str(self.client_address), os.path.abspath(dirname), self.change_dir_error))
+            return False
         data = os.listdir(dirname)
-        self.request.sendall(bytes('\n'.join(data), encoding='utf-8'))
+        if not data:
+            self.request.sendall(bytes('[]', encoding='utf-8'))
+        else:
+            self.request.sendall(bytes('\n'.join(data), encoding='utf-8'))
 
     def _del(self):
-        action, file_dir = self.recv.split()[0], self.recv.split()[1]
+        recv_data = self.recv.split()
+        action, file_dir = recv_data[0], recv_data[1]
+        if not check_root(self.basedir, os.path.abspath(file_dir)):
+            self.request.sendall(bytes('Can not change directory.', encoding='utf-8'))
+            self.logger.warning('[%s] Delete file "%s" failed, %s' % \
+                                (str(self.client_address), os.path.abspath(file_dir), self.change_dir_error))
+            return False
         if not os.path.exists(file_dir):
             self.request.sendall(bytes('File Not Found.', encoding='utf-8'))
+            self.logger.debug('[%s] Delete "%s" failed: file not found.' % \
+                              (str(self.client_address), file_dir))
             return False
         elif os.path.isfile(file_dir):
             os.remove(file_dir)
             self.request.sendall(bytes('Delete Success.', encoding='utf-8'))
+            self.logger.debug('[%s] Delete success: %s' % (str(self.client_address), file_dir))
         elif os.path.isdir(file_dir):
             shutil.rmtree(file_dir)
             self.request.sendall(bytes('Delete Success.', encoding='utf-8'))
+            self.logger.debug('[%s] Delete success: %s' % (str(self.client_address), file_dir))
 
     def _comm(self):
         try:
@@ -161,19 +206,18 @@ class MyTCPHandler(socketserver.BaseRequestHandler):
             self.request.sendall(bytes(data, encoding='utf-8'))
         except Exception as e:
             self.request.sendall(bytes(str(e), 'utf-8'))
-
+            self.logger.error('[%s] %s' % (str(self.client_address), e))
 
 
 def run(host, port):
     server = socketserver.ThreadingTCPServer((host, port), MyTCPHandler)
     with server:
         ip, addr = server.server_address
+        MyTCPHandler.logger.info('Start server. [%s:%s]' % (ip, addr))
         server.serve_forever()
 
 
 if __name__ == "__main__":
     host, port = 'localhost', 9999
     run(host, port)
-
-
